+++ src/Akeneo/Category/back/tests/Integration/Infrastructure/Storage/Save/Query/PurgeOrphanCategoriesSqlIntegration.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace AkeneoTest\Pim\Enrichment\Integration\Command\Category\RemoveCategoriesOrphan;
-
-use Akeneo\Category\back\tests\Integration\Helper\CategoryTestCase;
-use Akeneo\Category\Domain\Model\Enrichment\Category;
-use Akeneo\Category\Domain\Query\GetCategoryInterface;
-use Akeneo\Category\Domain\ValueObject\Code;
-use Akeneo\Category\Domain\ValueObject\PermissionCollection;
-use Symfony\Bundle\FrameworkBundle\Console\Application;
-use Symfony\Component\Console\Tester\CommandTester;
+declare(strict_types=1);
 
 /**
- * @copyright 2023 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @copyright 2023 Akeneo SAS (https://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class RemoveOrphanCategoriesIntegration extends CategoryTestCase
+
+namespace Akeneo\Category\back\tests\Integration\Infrastructure\Storage\Save\Query;
+
+use Akeneo\Category\back\tests\Integration\Helper\CategoryTestCase;
+use Akeneo\Category\Domain\Query\GetCategoryInterface;
+use Akeneo\Category\Infrastructure\Storage\Sql\PurgeOrphanCategoriesSql;
+
+class PurgeOrphanCategoriesSqlIntegration extends CategoryTestCase
 {
-    public function testCommandRemoveOrphanCategories(): void
+    public function testPurgeOrphanCategoriesFromDatabase(): void
     {
         $parent = $this->createOrUpdateCategory(
             code: 'parent',
@@ -49,16 +48,17 @@ class RemoveOrphanCategoriesIntegration extends CategoryTestCase
             rootId: $parent->getId()->getValue(),
         );
 
-        $output = $this->runRemoveOrphanCategoriesCommand();
-        $this->assertStringContainsString('0 rows removed', $output);
+        $affectedRows = $this->get(PurgeOrphanCategoriesSql::class)->execute();
+        $this->assertEquals(0, $affectedRows);
 
         $this->assertNotNull($child3->getParentId());
-        $this->setCategoryParentIdToNull('child3');
+
+        $this->setCategoryParentIdToNull((string) $child3->getCode());
         $child3 = $this->get(GetCategoryInterface::class)->byCode('child3');
         $this->assertNull($child3->getParentId());
 
-        $output = $this->runRemoveOrphanCategoriesCommand();
-        $this->assertStringContainsString('3 rows removed', $output);
+        $affectedRows = $this->get(PurgeOrphanCategoriesSql::class)->execute();
+        $this->assertEquals(3, $affectedRows);
 
         $nonOrphanCategoriesGenerator = $this->get(GetCategoryInterface::class)->byCodes(['parent', 'child1', 'child2', 'child3', 'child4', 'child5']);
         $nonOrphanCategories= [];
@@ -73,19 +73,6 @@ class RemoveOrphanCategoriesIntegration extends CategoryTestCase
 
         $this->assertCount( 3, $nonOrphanCategories);
         $this->assertEqualsCanonicalizing($expectedCategories, $nonOrphanCategories);
-    }
-
-    private function runRemoveOrphanCategoriesCommand(): string
-    {
-        $kernel = self::bootKernel();
-        $application = new Application($kernel);
-
-        $command = $application->find('akeneo:categories:remove-orphans');
-
-        $commandTester = new CommandTester($command);
-        $commandTester->execute(['command' => $command->getName()]);
-
-        return $commandTester->getDisplay();
     }
 
     private function setCategoryParentIdToNull(string $code): void